[log] Fix logging int16_ts

+++ src/core/include/tos/debug/sinks/sink.hpp
@@ -21,7 +21,7 @@ struct any_sink {
 
     template<class T = int16_t, typename = std::enable_if_t<!std::is_same<T, int>{}>>
     void add(int16_t i) {
-        add(static_cast<int16_t>(i));
+        add(static_cast<int64_t>(i));
     }
 
     template<class T = int32_t, typename = std::enable_if_t<!std::is_same<T, int>{}>>
